feat: :construction: Add methods to the dataTreatmentService.
Create the following public methods checkDrawn(call processDraw to increase the number of draws if true), checkVictories((call processVictory to increase the number of victories and processLoss to increase the number of losses)), countGoals (call processGoals to increase the goalsFavor and goalsOwn), calcEfficiency, calcTotalGames, calcTotalPoints, sortClassification, and the following private methods processVictory, processDraw, processGoals and processLoss. They will support the leaderboardService.